25 04 24 장바구니 수량 관련 유효성체크

+++ src/main/java/com/sixthsense/hexastay/service/StorecartService.java
@@ -16,7 +16,7 @@ import java.util.List;
 public interface StorecartService {
 
     /*1. 등록*/
-    Long addCart(StorecartitemDTO dto, Long hotelroomNum);
+    int addCart(StorecartitemDTO dto, Long hotelroomNum);
 
     /*2. 목록*/
     List<StorecartitemViewDTO> getCartList(Long hotelroomNum);
